Question u
Ah see I bought a refurbished phone and three network are saying due to a previous owner owing on a contract they had on the handset they won't unlock  so imei number has some sort of block although since I said this isn't right I bought in good faith and owe nothing personally a refurbished payasugo shouldn't b the case where I paid and have no signal where I live with their network .so every request was denied although I have since ranted and complained .so they said call again .so hopeful.this is why I wanted to root for opening whole new possibilities plus to enable me to unlock..was jellybean as stock since been updated by go mobile or previous owner then auto updated to kit kat. So I'm at a loss as to which direction to go.was an Xmas present and unable to use sufficiently for three mnths now .Any help or guide nice or suggestions as to where to go or how to go about be great fully appreciate  .I understand this isn't your problem or what ur service provides .so sorry for wasting any of your time.but as I'm here thought I'd ask a professional whom is skilled and trustworthy on a clearly trustworthy site.